+++ src/onegov.gazette/onegov/gazette/fields.py
@@ -1,26 +1,6 @@
-from bleach.sanitizer import Cleaner
-from onegov.form.fields import HtmlField as HtmlFieldBase
 from onegov.form.fields import MultiCheckboxField as MultiCheckboxFieldBase
 from wtforms import SelectField as SelectFieldBase
 from onegov.gazette import _
-
-
-cleaner = Cleaner(
-    tags=['br', 'em', 'p', 'strong', 'ol', 'ul', 'li'],
-    attributes={},
-    strip=True
-)
-
-
-class HtmlField(HtmlFieldBase):
-    """ A textfield with html with integrated sanitation.
-
-    We need a much stricter sanitation than the normal editor uses.
-
-    """
-
-    def pre_validate(self, form):
-        self.data = cleaner.clean(self.data)
 
 
 class SelectField(SelectFieldBase):