+++ src/onegov.gazette/onegov/gazette/tests/test_fields.py
@@ -1,36 +1,6 @@
 from onegov.form import Form
-from onegov.gazette.fields import HtmlField
 from onegov.gazette.fields import MultiCheckboxField
 from onegov.gazette.fields import SelectField
-
-
-def test_html_field():
-    form = Form()
-    field = HtmlField()
-    field = field.bind(form, 'html')
-
-    field.data = ''
-    assert 'class="editor"' in field()
-    assert field.validate(form)
-
-    field.data = (
-        '<h2><span class="">Lorem Ipsum</span></h2>'
-        '<p>'
-        '<span class="md-line md-end-block">'
-        '<span class=""><em>Lorem</em></span>, <a href="xx">ipsum.</a>'
-        '</span>'
-        '<span class="md-line md-end-block"><script>XXXX</script></span>'
-        '</p>'
-        '<ul><li>1</li><li>2</li><li>3</li></ul>'
-        '<ol><li>1</li><li>2</li><li>3</li></ol>'
-    )
-
-    field.validate(form)
-    assert field.data == (
-        'Lorem Ipsum<p><em>Lorem</em>, ipsum.XXXX</p>'
-        '<ul><li>1</li><li>2</li><li>3</li></ul>'
-        '<ol><li>1</li><li>2</li><li>3</li></ol>'
-    )
 
 
 def test_select_field():